[v1.6.4-rhel] [DO NOT MERGE] podman stats: calc CPU percentage correctly
When you run podman stats, the first interval always shows the wrong cpu
usage. To calculate cpu percentage we get the cpu time from the cgroup
and compare this against the system time between two stats. Since the
first time we do not have a previous stats an empty struct is used
instead. Thus we do not use the actual running time of the container but
the current unix timestamp (time since Jan 1 1970).

To fix this we make sure that the previous stats time is set to the
container start time, when it is empty.

[NO NEW TESTS NEEDED] No idea how I could create a test which would have
a predictable cpu usage.

See the linked bugzilla for a reproducer.

This was fixed with a cherry-pick in Podman v3.0.1-rhel with this commit:
https://github.com/containers/podman/pull/18710/commits/ca3d490f01b20a0e83b89cb57ff85245913598a7

However, the two files under the /pkg directory from that commit don't exist
in 1.6.4-rhel.  I've made the changes I think need to happen, but get a
segv almost immediately when running anything.  Putting this PR up in
case others spot the error.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2210140

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
Signed-off-by: Tom Sweeney <tsweeney@redhat.com>
Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>