podman machine init: add a --with-foreign-arch flag
The flag is qemu-specific.  It causes `qemu-static` & friends to be installed on the machine so that, e.g., x86_64 binaries can be run on aarch64 (aka "arm64").

Signed-off-by: Evgeny ("Zhenya") Roubinchtein <zhenya1007@gmail.com>

I have done manual testing as described in the PR message, so (I believe it is the case that) [NO NEW TESTS NEEDED]