Cirrus: Fix unsupported cirrus-cron build status
Every weekday when the `check_cirrus_cron` github-actions
workflow runs. It checks the status of all cirrus-cron jobs.  If a build
is found with a 'FAILED' status, it triggers an alert e-mail to be sent.
However, the `test_image_build` is marked as a manually-triggered,
resulting in a perpetual status of 'EXECUTING', even if there were
failures.  Fix this by only allowing the problematic task to run in pull
requests without the `[CI:DOCS]` magic keyword.

Signed-off-by: Chris Evich <cevich@redhat.com>