Cirrus: Use containers for more tasks
Containers are vastly faster to create and use, since they don't need to
"boot".  The automation_images repo. ensures the images are
substantially similar to the VMs in terms of packages and setup.
Previously, these container images could not be widely used in CI
because Cirrus-CI does not support sharing task-cache between VMs
& containers. Artifacts & curl don't have this limitation, allowing
wider use of containers in podman CI.

Signed-off-by: Chris Evich <cevich@redhat.com>